requirements.txt
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.2.0

